learning-python-for-forensics/Chapter 13/requirements.txt
xlsxwriter
colorama
python-registry == 1.1.0
pyfiglet
mutagen
pillow
simplekml
tqdm
unicodecsv
jinja2
lxml
